+++ src/main/java/com/course/admin/entity/Admin.java
@@ -1,32 +1,42 @@
 package com.course.admin.entity;
 
-import javax.persistence.*;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Table;
 
 /**
- * Created by 84074 on 2017/10/19.
+ * Created by Yan on 2017/10/24.
  */
 @Entity
-@Table(name="USER")
-public class UserEntity {
+@Table(name="ADMINISTRATOR")
+public class Admin {
 
     @GeneratedValue(strategy = GenerationType.AUTO)
-    @Id
+    @javax.persistence.Id
     private int id;
     private String username;
     private String password;
     private int roleId;
-    public UserEntity() {
+    //if roleId == 0, it is a librarian;
+    //if roleId == 1, it is a system administrator;
+
+    public Admin() {
 
     }
+
     public int getId() {
         return id;
     }
+
     public void setId(int id) {
         this.id = id;
     }
+
     public String getUsername() {
         return username;
     }
+
     public void setUsername(String username) {
         this.username = username;
     }
@@ -46,8 +56,4 @@ public class UserEntity {
     public void setRoleId(int roleId) {
         this.roleId = roleId;
     }
-
-
-
-
 }
