+++ src/main/java/com/course/admin/entity/Graduate.java
@@ -5,11 +5,11 @@ import javax.persistence.Entity;
 import javax.persistence.Table;
 
 @Entity
-@Table(name = "UNDERGRADUATE")
-public class UndergraduateEntity extends BorrowerEntity {
-
+@Table(name = "GRADUATE")
+public class Graduate extends Borrower {
     @Column
     private String major;
+    private String director;
 
     public String getMajor() {
         return major;
@@ -17,5 +17,13 @@ public class UndergraduateEntity extends BorrowerEntity {
 
     public void setMajor(String major) {
         this.major = major;
+    }
+
+    public String getDirector() {
+        return director;
+    }
+
+    public void setDirector(String director) {
+        this.director = director;
     }
 }
