+++ src/main/resources/application.yml
@@ -4,8 +4,8 @@ spring:
   datasource:
     driver-class-name: com.mysql.jdbc.Driver
     url: jdbc:mysql://127.0.0.1:3306/boot?useUnicode=true&characterEncoding=utf-8&useSSL=false
-    username: username
-    password: password
+    username: root
+    password: 123456
     platform: mysql
     type: org.apache.tomcat.jdbc.pool.DataSource
     dbcp2:
